Fixed cart icon
Put the width of the img in css instead of html. Used flexbox to make search bar/cart go into the second line and center the logo.

Slight issues with the icon jumping into the next line (might be able to be fixed by adjusting media queries or asking Gary).